requirements.txt
mysql-connector-python~=8.0.26
pycryptodome~=3.10.1